move Operator.h to abstract directory
