+++ src/shogun/mathematics/graph/ops/abstract/Operator.h
